fix: cleaned up hdf5_utils to only work with local files
